Basic Licensing for POC (#269)

+++ testing/start-e2e.sh
@@ -56,11 +56,6 @@ if [ "$(basename $(pwd))" != "testing" ]; then
     exit 1
 fi
 
-if ! docker build --no-cache --file ../Dockerfile --tag uh-cluster:testing ..; then
-    echo "docker build failed" 1>&2
-    exit 1
-fi
-
 if [ ! -d "$venv_dir" ] || [ "$venv_dir" -ot "$requirements_file" ]; then
     echo "Creating virtual environment ..."
 
@@ -80,8 +75,14 @@ fi
 
 . "$venv_dir/bin/activate"
 
+export UH_LICENSE="$(cat $PWD/../data/licenses/UltiHash-Test-1GB.lic)"
 
 if [ -z "$cluster_url" ]; then
+    if ! docker build --no-cache --file ../Dockerfile --tag uh-cluster:testing ..; then
+        echo "docker build failed" 1>&2
+        exit 1
+    fi
+
     docker compose up --detach
     trap "docker-compose down" SIGHUP SIGINT SIGQUIT SIGABRT EXIT
     cluster_url="http://localhost:8080"
